pyproject.toml
[project]
name = "psychology-club"
version = "0.1.0"
description = ""
authors = [
    {name = "katarzynamariaelzbieta",email = "kasia.gajc@wp.pl"}
]
readme = "README.md"
package-mode = false

requires-python = ">=3.11,<4.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-security-too (>=5.6.2,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "flask-mailman (>=1.1.1,<2.0.0)",
    "dash (>=3.2.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "dash-mantine-components (>=2.3.0,<3.0.0)",
    "dash-iconify (>=0.1.2,<0.2.0)",
    "bleach (>=6.2.0,<7.0.0)",
    "dash-extensions (>=2.0.4,<3.0.0)",
    "tinycss2 (>=1.4.0,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "dash-ag-grid (>=33.3.3,<34.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
